Pop ups added

+++ lib/screens/home/screen3/message_content.dart
@@ -2,6 +2,7 @@ import 'package:brew_crew/screens/home/screen3/editmessage.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import '../homescreen.dart';
+import 'package:flutter/cupertino.dart';
 
 class MessageContentPage extends StatelessWidget {
   MessageContentPage({super.key, required this.message, required this.title});
@@ -37,37 +38,36 @@ class MessageContentPage extends StatelessWidget {
             width: 10,
           ),
         ],
-        
       ),
-        body: Padding(
-          padding: EdgeInsets.fromLTRB(15, 30, 15, 0),
-          child: Container(
-            padding: EdgeInsets.all(20),
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(12),
-              color: Colors.white,
-              boxShadow: [
-                BoxShadow(
-                  // color: const Color.fromRGBO(
-                  //         0, 0, 0, 0)
-                  //     .withOpacity(0.04),
-                  color: const Color.fromRGBO(50, 50, 93, 0.25).withOpacity(0.08),
-                  // color:
-                  //     const Color.fromRGBO(50, 50, 93, 0.25).withOpacity(0.1),
-                  spreadRadius: 10,
-                  blurRadius: 20,
-                  offset: const Offset(0, 8), // changes position of shadow
-                ),
-              ],
-            ),
-            child: Text(
-              message,
-              style: TextStyle(
-                fontFamily: "Montserrat",
+      body: Padding(
+        padding: EdgeInsets.fromLTRB(15, 30, 15, 0),
+        child: Container(
+          padding: EdgeInsets.all(20),
+          decoration: BoxDecoration(
+            borderRadius: BorderRadius.circular(12),
+            color: Colors.white,
+            boxShadow: [
+              BoxShadow(
+                // color: const Color.fromRGBO(
+                //         0, 0, 0, 0)
+                //     .withOpacity(0.04),
+                color: const Color.fromRGBO(50, 50, 93, 0.25).withOpacity(0.08),
+                // color:
+                //     const Color.fromRGBO(50, 50, 93, 0.25).withOpacity(0.1),
+                spreadRadius: 10,
+                blurRadius: 20,
+                offset: const Offset(0, 8), // changes position of shadow
               ),
+            ],
+          ),
+          child: Text(
+            message,
+            style: TextStyle(
+              fontFamily: "Montserrat",
             ),
           ),
         ),
+      ),
     );
   }
 }
@@ -127,13 +127,52 @@ void bottomsheet(BuildContext context, String message, String title) {
                 onTap: () {},
                 child: ListTile(
                   onTap: () {
-                    DocumentReference _docref = FirebaseFirestore.instance
-                        .collection("message")
-                        .doc(title);
-                    _docref.delete();
-                    print('${title} deleted');
-                    Navigator.pop(context);
-                    Navigator.pop(context);
+                    // DocumentReference _docref = FirebaseFirestore.instance
+                    //     .collection("message")
+                    //     .doc(title);
+                    // _docref.delete();
+                    // print('${title} deleted');
+                    // Navigator.pop(context);
+                    // Navigator.pop(context);
+                    showDialog(
+                      context: context,
+                      builder: (BuildContext context) => CupertinoAlertDialog(
+                        content: const Text(
+                          'Are you sure you want to delete this message?',
+                          style: TextStyle(
+                            fontFamily: "Montserrat",
+                          ),
+                        ),
+                        actions: [
+                          CupertinoDialogAction(
+                            child: const Text(
+                              'Yes',
+                              style: TextStyle(color: const Color(0XffA85CF9)),
+                            ),
+                            onPressed: () {
+                              DocumentReference _docref = FirebaseFirestore
+                                  .instance
+                                  .collection("message")
+                                  .doc(title);
+                              _docref.delete();
+                              print('${title} deleted');
+                              Navigator.pop(context);
+                              Navigator.pop(context);
+                              Navigator.of(context).pop();
+                            },
+                          ),
+                          CupertinoDialogAction(
+                            child: const Text('No',
+                                style: TextStyle(
+                                    fontFamily: "Montserrat",
+                                    color: const Color(0XffA85CF9))),
+                            onPressed: () {
+                              Navigator.of(context).pop();
+                            },
+                          ),
+                        ],
+                      ),
+                    );
                   },
                   dense: true,
                   title: Row(
